+++ tests/integration/test_compile.py
@@ -89,7 +89,7 @@ def test_compile(fake_model_training_session, device, mode):
         # test both devices if possible, where we are testing swapping of devices between compile and
         # loading (CUDA -> cpu and vice versa)
         chemical_symbols = config.chemical_symbols
-        ckpt_calc = NequIPCalculator.from_checkpoint_model(
+        ckpt_calc = NequIPCalculator._from_checkpoint_model(
             ckpt_path,
             device=load_device,
             chemical_symbols=chemical_symbols,