+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Most recent change on the top.
 ## Unreleased
 
 
+### Changed
+- [Breaking] Hide `NequIPCalculator.from_checkpoint_model` and `NequIPCalculator.from_packaged_model` methods by making them private; users should only ever use `NequIPCalculator.from_compiled_model`
+
+
 ## [0.10.0]
 
 ### Changed